requirements.txt
Flask==2.3.2
gunicorn==21.2.0
Pillow==10.0.0
pip==23.1.2
tensorflow==2.13.0
